requirements.txt
django == 1.8.6
django-geojson [field]==2.8.1
django-leaflet